Refactor: alterando o program.cs para salvar o valor em decimal no banco de dados e migrations da atualização da Model.

+++ WebPostgreSql/Models/Produto.cs
@@ -16,6 +16,6 @@ namespace WebPostgreSql.Models
 
         [Column("Valor")]
         [Display(Name = "Valor")]
-        public float Valor { get; set; }
+        public decimal Valor { get; set; } // Alterado para decimal
     }
 }
